requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.12.1
bleach==6.0.0
blobfile==2.0.1
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.1
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
deepspeed==0.8.3
defusedxml==0.7.1
dill==0.3.6
Django==4.2
docopt==0.6.2
einops==0.6.0
evaluate==0.4.0
exceptiongroup==1.1.1
fairscale==0.4.0
fastjsonschema==2.16.3
filelock==3.11.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grad-cam==1.4.8
graphviz==0.20.1
grpcio==1.54.0
hjson==3.1.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.33.1
iniconfig==2.0.0
install==1.3.5
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.0
loralib==0.1.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mistune==2.0.5
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy==1.2.0
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open-clip-torch==2.23.0
opencv-python==4.7.0.72
pandas==2.0.0
pandocfilters==1.5.0
pathspec==0.12.1
Pillow==9.5.0
pipreqs==0.4.13
pluggy==1.0.0
prometheus-client==0.16.0
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pycryptodomex==3.17
pydantic==1.7.3
pydevd==2.9.6
pydicom==2.3.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.97
SimpleITK==2.3.1
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.12.9
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
torchviz==0.0.2
tornado==6.3
tqdm==4.65.0
transformers==4.32.1
triton==2.0.0
ttach==0.0.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
